refactor: rename `change_hexrays_config` to `modify_decompiler_config` and make it `&mut`
